RunGen: bounds-query failure shouldn't matter if we use estimates anyway
bounds_query_input_shapes() agressively fails if we can't complete the bounds-query, which can happen if the constraints on inputs are nontrivial. While we can (and should) improve the bounds-query logic to make this more robust, we shouldn't aggressively fail here in the first place, as the bounds-query shape(s) end up unused if we have estimates for the inputs (which we usually do). This just adds a ShapePromise type that wraps access to the resulting Shapes inside a function; if we never need the shape, the failure doesn't matter and never happens.