Clean up default autotuner output
This restructures so that the default output of the autoscheduler (at HL_DEBUG_AUTOSCHEDULE=0) is much smaller, with just some output for each pass. (All the previous output is still present for HL_DEBUG_AUTOSCHEDULE=1.) Not entire sure if this is the best "per-pass" summary; definitely open for suggestions here (even shorter/terser would be even nicer here IMHO).

Also, a little drive-by cleanup to the progress bar (to avoid leaving them dangling) and to give slightly better feedback during the autotune script.