Allow Target::has_feature() to operator on halide_target_feature_t
This is really a workaround to allow generated .schedule.h files to compile happily in all cases. A better long-term answer would be for those files to use the Target::Feature names (rather than the halide_target_feature_t names), but doing so will require an additional value->name map (which must be maintained and could get out of sync), so I'm reluctant to do that without a little more thought.