New select rewrites for simplifying unrolled slid loops
Unrolling happens after sliding but before partitioning, so the extents
of the unrolled loop should have an upper bound, but it can be hard for
the compiler to see this through all the selects. This PR makes things
way more aggressive in the about-to-fail case in unrolling. It now
exhaustively substitutes in everything for maximally aggressive
simplification, and there are a bunch of new simplifier rules to help
cancel things through select nodes. This exhaustive substitution is
somewhat dangerous, but this is a case where the compiler is about to
throw an error if it doesn't simplify down to almost a constant.