Remove HL_WEIGHTS_OUT_DIR
It was only referenced in single-shot scheduling, but we shouldn't ever be saving from that path anyway (only from the retraining loop).

Also change save_weights() to fail with abort() if you try to save with an empty output path (rather than quietly doing nothing), and upgrade save failures from assert(0) to abort(), because that's not an error we want to risk eliding based on compiler settings.