Remove min_iter and max_iter from benchmark & RunGen
These aren't really that useful; min_time is usually the only knob of this sort that is interesting to fiddle with. Remove them entirely to simplify code and reduce the temptation to mess with them. (Some versions of gbenchmark don't support min_iter and max_iter at all, only min_time.)