Split FunctionDAG into .h and .cpp files
The driving force for this change is to move the helper funcs for `Autoscheduler::BoundContents::ref_count` out of the header file: the current definition can cause possible ODR violations as-is; we could avoid this by marking them `inline`, which would likely cause undesirable code replication.

Moved most of the larger function bodies from the .h to the .cpp file, but otherwise not much in the way of refactoring.