Ensure that ~ErrorReport is marked as noreturn
C++ wasn't aware that ~ErrorReport() never returns; as a result, clang-tidy can issue complaints about e.g. "possible null pointer access" that aren't possible because the code is guarded by `internal_assert` or similar. This PR adds the C++11 [[noreturn]] annotation as needed to ensure that the compiler does in fact know this; whether this has meaingful impact on code quality isn't clear, but at a minimum, it allows us to enable clang-tidy checks that previously would have sprayed too many false-positive warnings to be usable.

(Note of course that the 'never returns' doesn't apply to warnings; to allow for this, I split the functionality into ErrorReport and WarningReport accordingly.)