Change Python extension from .py.c -> .py.cpp
In theory, the Python extension file we generate can be compiled just fine as a plain-C file... but if we are building with cpp-name-mangling enabled in the target, we will include *generated* .h files that can't be compiled as plain C.

We really don't want to vary the file extensions based on target flags, and in practice, it's extremely unlikely that anyone needs to rely on this being pure C output (vs possibly C++). So just change the default.