Remove the -x option from Generators
This was used to remap the default filename extensions for various outputs, but was really being used as a workaround for defaults that weren't good defaults. Those defaults have mostly been changed (except for .s, which isn't relevant here), so we should simplify the code by just removing this as an unnecessary complication.