Clean IRPrinter indentation code
It's always bugged me that IRPrinter used a separate call (`do_indent()`) to indent outputs, rather than some approach that mixed into the stream flow of operator<<. This is a completely nonessential change that nevertheless I think improves the code; it creates a new type, `Indentation`, which has an `operator<<` overload. This means that stream writes that begin with an indent are more obvious (since it's part of the ordinary stream syntax).