CMake: add optional HALIDE_SHARED_LLVM_LIBRARY option.
As discussed in https://github.com/halide/Halide/issues/4284
one important thing is to be able to link to shared versions
of LLVM libraries. This is off-by-default.

The important [existing] caveat to note is that i don't see
where those llvm libraries-to-be-linked are specified anywhere
in the halide_config.cmake, so as long as you are
a. building static halide
and
b. not building halide as a cmake sub-project (via add_subdirectory)
i don't see how the linking can succeed, both either with
old static llvm libraries, and now with shared llvm library.

The best i can tell `llvm-config --system-libs` is only needed
when linking to static libraries, to propagate their deps,
so let's not do that for shared llvm libs.

And the switch to `llvm_config()` is the main thing here,
it internally handles `USE_SHARED`, and does `target_link_libraries()`.

Since this is opt-in, i don't believe this should have any impact
on existing users.