Link libHalide for building libauto_schedule.so
Passing libHalide.so as an input while building the autoscheduler
inside apps/autoscheduler is creating a dependency on libHalide for
the shared plugin library where libauto_schedule.so looks for
libHalide.so in the absolute path and fails to open if the libHalide
is not found in that path. This makes the plugin not portable, as it
warrants the libHalide.so to be present in the same location while using
the plugin.

Using readelf -d on the built libauto_schedule.so --

`(NEEDED)             Shared library: [/mnt/workspace/Halide/distrib/bin/libHalide.so]`

While using the plugin, if libhalide is loaded from other path, the
plugin fails to load as it searched for libHalide in the absolute
path used while building.

This change links to the built libHalide.so instead of passing it as
an input file while building experimental autoscheduler plugin
libauto_schedule.so.