Added library path to search for libHalide
Generators using the autoscheduler plugin need to specify
the path where the plugin searches for libHalide.so when
generator is executed, or the LD_LIBRARY_PATH should point
to it.

This change adds GENERATOR_LDFLAGS with the required linker
flags, to be used for building generators which use the
autoscheduler plugin.

This also removes the libHalide link added for building
libauto_schedule.so by the previous commit as shared libraries
don't need to be linked with their dependencies on build-time.