Drop support for Python 2.x
Most of the world is dropping support for Python 2.x no later than January 1, 2020; in theory our Python bindings still support 2.x, but in practice we haven't been reliably testing that configuration. Let's go ahead and drop any pretense of support now; this removes the last dregs of code that assumes that Python 2.x is a possibility.