Fix offload mode autoscheduler results.
Currently, when autoscheduler is invoked in offload mode, the
the emit options {featurization and schedule} are not working
as expected and bailing out due to auto_scheduler_results not
present in the lowered module, resulting in empty feature and
schedule files produced. This happens because, when offloaded
submodules are recursively resolved, the autoscheduler results
are not copied back into the lowered module. This change fixes
this issue by setting back the auto_scheduler_results in the
ModuleContents of the module that is lowered after resolving
submodules.