Define div/mod by zero to return zero
Also define things like -128/-1 to just wrap, for types that wrap.

This makes div/mod side-effect-free, which removes the need for
quiet_div/quiet_mod, and avoids runtime faults due to scheduling (see

Not yet ready to merge, as the simplifier rules need to be reverified
under this new constraint (there may now be division by zero on the LHS,
where previously we occasionally assumed that there was not, because it
was UB).