Increase autograd estimates/sizes from 32 -> 64
The classic autoscheduler can over-vectorize (e.g., on Skylake it will vectorize to width=64, which will fail for inputs of width<64). This is a somewhat-known issue; rather than try to fix it here, I am just going to widen the buffers to the autograd test to avoid failure on Skylake+ machines. (Note that autograd shouldn't be using the classic autoscheduler in the first place; it should be using the gradient autoscheduler, but packaging issues make that impractical just yet.)