MSAN: check input buffers for poison (Issue #4481)
Augment our MSAN support to ensure that input buffers (and the output of extern stages) are checked for uninitialized data.

Note 1: msan_aottest has become an ugly mess of code; it should probably be changed to use a table-driven state machine. Not sure if that blocks this or not.

Note 2: this should probably get checked inside Google before landing (which will probably happen post-holidays), as we have a fair number of complex pipelines that are running with msan enabled.