Permit dynamic allocations in kernels
and pass the peak amount of heap memory we're going to use to
halide_{cuda/opencl/metal/etc}_run so that we tell the API to allocate
some heap.

This is a WIP. It currently does nothing for non-CUDA, and on CUDA it
codegens a dynamic allocation (which should be a call to malloc) as just
a null pointer.