Make the Simplifier handle PureExtern functions via LUT
There were a number of PureExtern functions we didn't handle (notably sin and cos, but many others too); rather than add even more if-else clauses, I refactored the PureExtern case into a set of LUTs to reduce redundant code.

Note that the actual motivation for this change was to ensure that sin(CONSTANT) and cos(CONSTANT) got simplified to a real constant (rather than putting us at the mercy of LLVM).

(With more cleverness we could make it a single LUT, but my first approach involved template-fu that was overly complex to read; I suspect that this approach is a better balance of compile time vs code complexity, but would welcome someone more clever to simplify it further.)