Tweak relative_error in math.cpp (Issue #4559)
Changes to LLVM's math reassociation made our `exp()` implementation slightly less accurate in some float32 cases than they already were... but not by much, so this is really just a tweak for reality.

e.g., `exp(18.125)` should be ~74402492.45; calling `std::exp()` returns 74402496.0  (on the systems I checked), for an absolute error of 4.

Prior to the LLVM change, our non-FMA code on x86 would reliably produce 74402552 (abserr: 60, relerr: ~0.806e-6); now, it produces 74402416.0 (abserr: -76, relerr: ~1.07e-06), just slightly outside our bounds.

Trying to maintain the same result in the face of compiler changes and different instruction sets is a fool's errand, and the number of values that are "worse" is small, so just bumping the error value is a fair fix.

(Though it's worth pointing out that the 'reference' value in this case is actually inaccurate; we might be better off always calculating the reference value as a double and coercing to a float.)