Add host_supports_device_api()
This adds a new call, `host_supports_device_api()`, which is intended to allow JIT callers to determine whether a given DeviceAPI is likely supported on the current host. Modified tutorial code to use it as appropriate. (Note that existing logic in e.g. lesson_12 that tries to load driver libraries isn't adequate, as a system may have the right software installed but no suitable hardware.)