Change PyBind11 build rules to rely on a pip-installed version
Up to now, we've required people to explicitly clone PyBind11 and point the build files at it via PYBIND11_PATH; as it turns out, there is a pip package for it, which is much simpler to manage.

This change removes usage of PYBIND11_PATH entirely, and assumes that there is a pip install that has all includes set up properly (which should be the case if you have used requirements.txt).

If this lands, there will be a followup to remove the explicit PYBIND11_PATH cruft from the buildbots.