Increase stack size for wasm tests as needed
CodegenLLVM deliberately makes many small (register-sized) calls to alloca, on the plausible assumption that CPU-ish backends will promote these into registers. This isn't true for the WebAssembly backend, since (as a stack machine) it doesn't do registers; in degenerate cases (e.g. a func with ~everything inlined, and tracing enabled to increase stack usage), we can overflow the default wasm stack of ~64k, which crashes in amusing ways.

To mitigate this, add some code to CodegenLLVM to track the total alloca usage, then have WasmExecutor add this amount to the stack-size specified at link time.

Note 1: AFAICT there isn't a pre-baked LLVM method for "how much stack space does this Module require"; if there is a cleaner way to do this without tracking it ourself, that would be great.

Note 2: would probably also be profitable to see if we can avoid the 'assume alloca -> register' design for the wasm backend; even if we do that, this fix is probably still desirable for other possible degenerate cases of wasm.

Note 3: drive-by fix in WasmExecutor for LLVM >= 11.0 error case of linker.