Fix incorrect int-promotion in Python bindings
Fixes a subtle bug in the python bindings regarding int and float handling in binary ops.

(1) In Halide, a binary op with an integer literal and an Expr should implicitly coerce the integer literal to the same type as the Expr; in Python, this was handled correctly for the cases where one type was *really* an Expr, but not for the cases where it was merely Expr-like (e.g. a FuncRef). This cleans that up and adds tests.

(2) Adding the fix above ended up breaking our detection for double->rounding errors, because we only handled this for explicit Expr conversions, rather than when the conversion was implicit from a binary operator. Restructured the checking for this case so that it should be more uniform.