turn on clang-tidy's modernize-pass-by-value
In constructors that take a copy of an argument, pass by value and
std::move instead of passing by const ref and making a copy. This is
strictly better for two reasons.

1) This avoids a copy in the case where the argument was an rval.
2) The type signature loses some line noise, and makes it more obvious
to callers that a copy is going to happen.

There were a healthy number of cases this revealed where the object
should *not* have been making a copy of the argument. These were largely
local IRVisitors and IRMutators doing some sort of tightly-scoped
analysis or mutation. so I changed those to have const ref members and
continue to take the argument by const ref.

Also added a makefile target to get clang-tidy to automatically apply
suggested changes.