Fix invalid extern stage
Bounds queries are supposed to nest. If a region O of the output
requires region I of the input, then a subset of O is supposed to
require a subset of I. Not sure what was going on with this extern
stage, but it didn't obey that property. I rewrote it.