Add a promise_clamped intrinsic
This intrinsic has proved useful in two branches as a better way to
lower GuardWithIf. Instead of relying on bounds inference being able to
understand the if statement, we just declare the bounds of the variable
inside the if using a version of clamp that is guaranteed to be a no-op
at the place it occurs.

This PR adds the intrinsic and simplification rules for it.