Add simplifier rules for for/if combinations
First, in the following:
```
if (extent > 0) { for (x = 0; x < extent; x++) {...}}

```
The if is redundant and can be removed.

Next, we can fold away `if (likely(true/false))` in the same way we fold
away `if (true/false)`

Finally, we used to transform:
```
for (x = 0; x < min(1, foo); x++)
```
to
```
if (0 < foo)
```
Because a for loop where the extent is at most one is just an if.

However, this is not a good mutation if done in a vectorized loop,
because a for loop with an extent that varies across the vector lanes is
considered a likely-if, and an all-true vectorized case is generated.
This doesn't happen for the regular if. So I just squashed this
transformation if we're inside a vector loop.