Replace stream << 'c' with stream << "c"
This is wild, but both libc++ and libstdc++ inline operator<<(char) to
something that takes the char by address, so it gets pushed onto the
stack before the call. The const char * version is just a global
constant with a constant address that can be passed in a register.

Making this change reduces the size of libHalide.so by epsilon (a single
4k page), but I claim it's a good idea because we currently use a mix of
<< 'c' and << "c", and this makes it uniform.

Also replaced std::endl with "\n" in some places. endl is for when you
*also* want to flush the output stream. This is usually when you want to
send a message to std::cout and you anticipate crashing shortly
afterwards. E.g. "first test passed, here comes the second one...", or
"running test with seed <...>".

Also found some cases of foo << "string2" << "string2"

Finally, in LLVM_Output.cpp I replaced out << '\0' with out.put('\0'),
to make it super-explicit that a single null character was being pushed.
Clearly out << "\0" would not have done the right thing. Did something
similar for another mysterious hex-code char too.