Move first simplification before allocation bounds inference
Removed the dependence of allocation bounds inference and storage
folding (and boxes_touched) on magic names, which means we can run full
simplification earlier. This should generally make allocation bounds
tighter, and also speeds up the compiler by giving allocation bounds
inference less IR to chew through (by a full 10% on local laplacian
lowering, and a similar amount on a 32x32 fft, 5% on lens blur).

See #4713 for the previous version of this PR.