Simplify min(likely(x), x) to x
We were simplifying it to likely(x), which I believe is wrong. A likely
marks which branch of the immediately containing min/max should be
optimized for. Once it has been optimized for that, the likely should
vanish. It's not supposed to be transitive to containing min/max nodes
too.