Teach lower warp shuffles to recognize lane <= 0
The warp shuffle pass needs to know when code is dropping down to a
reduced number of threads. Previously it only understood lane <
constant. This teaches it to recognize the lane <= constant too.

Unfortunately I don't currently have a failing case that this fixes, but
I did during an intermediate state of another branch.