Generalize uniquify_variable_names
Currently uniquify_variable_names only works on exactly the kind of
duplicate names we tend to generate in bounds inference, and only if you
run it on the entire stmt. In other circumstances it does odd things
(e.g. attach negative integer suffixes to free variables!), and we risk
introducing bugs, as I discovered in another PR.

This PR fixes uniquify_variable_names to do the right thing in the
general case, and adds tests.

Also added some needed functionality to Scope - you can now query the
number of nested definitions of the same thing, and you can move a
Scope.