Rearrange lowering to fix input bounds inference issue
add_image_checks needs to happen after bounds inference so that it
detects buffer overreads on the input. In its current place it only
detects if the input is large enough for the output to be correct, but
it doesn't check for the presence of guard bands to cover loading
out-of-bounds values that then get discarded. In theory this could
fault, if a guard band is needed on a dimension with large stride.

For this to not break a whole bunch of code due to false positives with
overly strict bounds, we need better simplification. So the first
simplification pass is moved before allocation bounds inference. As a
side benefit, this also reduces the amount of IR that pass needs to chew
through, so it speeds up lowering.

Fixes #3036