Reduce number of copies in Scope::push
pushing something onto a scope copies it, so if we accept rvals
appropriately then we can avoid those copies. Also makes it possible to
use scopes with non-copyable types as values. Speeds up the compiler
about half a percent on local laplacian. Increases the size of libHalide
by two pages (8k), so I guess the speed-up is due to inlining.