Permit a new class of benign race condition
It's safe(*) to parallelize/split/reorder a reduction domain if all
possible races are races to store the same value to the same index. For
example:
```
f(g(r)) = 4;
```

It's also okay to use the pure vars on the RHS, because for any possible
write-after-write hazard we know the value of the pure var is the same
because we require the pure var to appear unadorned on the LHS:
```
f(g(r, x), x) = h(x);
```

(*) In C/C++ a "benign race condition" is a contradiction in terms,
because it's UB. I'm think LLVM IR makes similar assumptions in its
memory model. The closest example I could find to this race condition
that has a concrete failure mode in C/C++ is section 2.4 in
https://www.usenix.org/legacy/event/hotpar11/tech/final_files/Boehm.pdf.
However that example isn't a straight-up write-after-write hazard where
the writes have the same value. There are also unsynchronized reads
involved, and the reads are essential to get a failure (how else could
you observe one?). This PR explicitly prohibits cases that read the
values being written to. No reads can occur until the parallel loop
ends, which acts as a synchronization point. I'm not going to claim that
a failure is impossible, but I can't think of one. If we manage to
construct a failure for this kind of race we'll have to revisit.

Note that Halide currently permits this kind of race condition in
another context: We redundantly re-evaluate points near the end of a
split domain with shiftinwards, even when the outer loop is
parallelized:

```
f(x) = h(x);
f.split(x, xo, xi, 8, TailStrategy::ShiftInwards).parallel(xo);
```

In practice this sort of race is *not* benign with floating point
right-hand-sides, as the parallel loop could be partitioned and compiled
as two separate racing pieces of code, which undergo different floating
point optimizations, then the values we're racing to store could then
actually be different. This often manifests as a strip of
nondeterministic values near the bottom or right edge of an image in
algorithms that are sensitive to floating point nonsense. The
strict-float target flag is useful for people who worry about such
things. Or just don't write schedules with this class of race condition
in them. We try to not let users shoot themself in the foot, but let's
not put them in a padded cell.

Fixes #4658