fix build when -stdlib=libc++ was passed into the plain C test
When LLVM was built with it, the -stdlib=libc++ flag would get passed to
correctness_plain_c_includes, which is (correctly) only using the clang
C compiler driver, not clang++, where this is an invalid option.

Since CMake doesn't seem to have per-language options apart from
CMAKE_CXX_FLAGS which people discourage from directly modifying,
we use a conditional instead of the string itself.

Note that passing the option to the linker is fine as is.