Delete six dubious simplifier rules
The deleted rules try to make negative constants positive, but don't
reduce the total op count, and replace an addition with a subtraction,
which doesn't count as a strength reduction.

Making negative constants positive may be defensible as a strength
reduction, but other rules may make positive constants negative again,
and elsewhere we treat replacing subtraction with addition as a strength
reduction. So we were unable to prove that these rules don't cause
cycles with these other rules (and hence infinite recursion). It would
be good to just delete them if possible.

In the open source tests, this required adding a few more variants of an
EQ rule on selects. These were formally verified. In the strength
ordering the new rules are cycle-free, even though they increase the op
count in two cases. They reduce the total number of leaf nodes by one
(c0 goes away) and we're treating the leaf count as higher priority than
the op count in our total ordering over expressions.