Internal compiler code can introduce shifts by greater than the type size.
So this should be an overflow IR node instead of being an eager user
assert.

The specific case was a can_prove(x << y) substituting in random values
for y when HL_DEBUG_CODEGEN>0 to probe for counterexamples. Some of
these random values where greater than the bit-size of x.