Don't automatically fold storage along vectorized dimensions
It's generally a bad idea because loads and stores scalarize, but
there's no way in the schedule to turn it off, so you can't write an
optimal schedule for the case in the newly added warning test where you
want sliding but not folding. This PR makes folding opt-in in this case
instead.

When automatic folding is rejected due to this test, we emit a warning
that includes the schedule to use to explicitly request the folding that
used to occur.

Longer term it would be good to disable automatic folding entirely. This
PR can be seen as a trial of that in one specific case.