Keep constants outermost in LICM
LICM is prone to making a bit of a mess. I found it can be tamed by not
letting it sink constants to the bottom of an expression, and by running
CSE before grouping loop invariants instead of after to stop it from
reassociating things in way that break up common subexpressions.

This provides small speedups on several of the apps. The largest gain is
1-2% on local_laplacian.