Restore division rule
PR #4896 reduced a rule that a term from the numerator that was a
multiple of the denominator to only triggering in the case where it was
equal to the denominator. This was done to respect the reduction order
and keep the simplifier cycle free. We forgot about the case where the
term is multiplied by the negative of the denominator (e.g. (x*-2 -
y)/2). This can also be pulled out without breaking the reduction order,
because the multiply goes away.