cmake: llvm: fix linking against LLVM shared lib
Fixes the build with -DLLVM_USE_SHARED_LLVM_LIBRARY=ON to
actually use the dynamic build (libLLVM.so). Relevant
for LLVM builds with -DLLVM_BUILD_LLVM_DYLIB=ON. Tested
with LLVM v10.0.0.

There are two issues that prevent a build against LLVM shared lib:
1. cmake failure:

      CMake Error at /usr/lib/llvm/10/lib64/cmake/llvm/LLVM-Config.cmake:145 (message):
        Target NVPTX is not in the set of libraries.
      Call Stack (most recent call first):
        /usr/lib/llvm/10/lib64/cmake/llvm/LLVM-Config.cmake:270 (llvm_expand_pseudo_components)
        dependencies/llvm/CMakeLists.txt:141 (llvm_map_components_to_libnames)

2. missing -lLLVM on the link line

Regarding issue 2: the result of missing -lLLVM on the link line can
be either of the following scenarios both of which are wrong:
(A) (in theory) a successful build but with all LLVM modules linked into
libHalide.so, which is not what the user requested with
LLVM_USE_SHARED_LLVM_LIBRARY=ON,
(B) (observed) a broken libHalide.so, where linking of apps against it fails with
undefined symbols errors related to target components (related to issue
1, see more below):

      /usr/lib/gcc/x86_64-pc-linux-gnu/10.1.0/../../../../x86_64-pc-linux-gnu/bin/ld:
      src/libHalide.so: undefined reference to `LLVMInitializeX86Target'

Sidenote regarding (B), linking the app against -lLLVM makes the link
succeed but fails at runtime, because then some static LLVM data
is linked into the app binary twice, so components get loaded twice:

      : CommandLine Error: Option 'pm-max-devirt-iterations' registered more
      than once!
      LLVM ERROR: inconsistency in registered CommandLine options

The above errors from (B) are symptomps that are not important, the fix
is to just link libHalide.so against libLLVM.so.  This commit provides
two alternatives for achiving it:
(a) use llvm_config()
(b) woraround by manually doing what llvm_config() does

I think the correct way is (a), however it cannot be used yet, because
in lateset LLVM 10.0.0, llvm_config does not accept targets of
type INTERFACE_LIBRARY (see attached patch for what it would take
to add that support to LLVM). If/when LLVM starts supporting
INTERFACE_LIBRARY, then (b) should be deleted in favor of (a).

The problem with existing cmake code that tries to link against LLVM
library(ies) was that it seemed to communicate the intent to
LLVM-Config.cmake via LLVM_USE_SHARED var, but LLVM-Config doesn't use
that var. Instead, the llvm_config method takes an argument option
USE_SHARED.

Regarding issue 1: appears to be caused by treating targets as components.

Observed when LLVM is built into a shared library with:
	-DBUILD_SHARED_LIBS=OFF
	-DLLVM_BUILD_LLVM_DYLIB=ON
	-DLLVM_LINK_LLVM_DYLIB=ON
	-DLLVM_DISTRIBUTION_COMPONENTS="comp1;comp2;..."
	-DLLVM_TARGETS_TO_BUILD=""
	-DLLVM_EXPERIMENTAL_TARGETS_TO_BUILD="X86;NVPTX;...."

Note that LLVM_DYLIB_COMPONENTS is NOT set explicitly, and
I am not sure what exactly it defaults to, but the shared lib
(libLLVM.so does end up containing all components selected by
LLVM_DISTRIBUTION_COMPONENTS and all selected targets. (Not that it
matters, but this is the LLVM config used in the package recipe in the
Gentoo distribution.)

The selected LLVM components do get built as static archives (e.g.
LLVMSupport.a, etc.) and do get linked into libLLVM.so and do get
installed into the system, however, there the static archives
for target libraries (e.g. LLVMX86TargetInfo.a, etc) get linked but
do NOT get installed into the system.

So, when Halide cmake script asks for the target libraries as components
(in call to llvm_map_components_to_libnames), that call fails),
presumably because the corresponding static archives for the target
libraries do not exist in the system. However, when linking against
the LLVM shared library (as opposed to static archives), it is
not necessary to specify the targets as components on the link line.
The target components are linked into the shared object, along with the
rest of the components, and specifying just the shared object is sufficient.

Note that llvm_config implementation specifies both the shared object,
followed by static archives for sake of fallback, so I kept that same
logic. It would work equally well to only list LLVM without the static
objects for the case with LLVM_USE_SHARED_LLVM_LIBRARY=ON.

For reference, to enable approach (a), the patch to add support for
targets of type INTERFACE_LIBRARY in llvm_config() to LLVM v10.0.0:

    --- a/cmake/modules/LLVM-Config.cmake   2020-06-08 19:36:44.804248189 -0000
    +++ b/cmake/modules/LLVM-Config.cmake   2020-06-08 19:37:23.191439807 -0000
    @@ -87,7 +87,13 @@
           endif()
         endif()

    -    target_link_libraries(${executable} PRIVATE LLVM)
    +    get_target_property(t ${executable} TYPE)
    +    if(t STREQUAL "INTERFACE_LIBRARY")
    +      set(link_mode "INTERFACE")
    +    else()
    +      set(link_mode "PRIVATE")
    +    endif()
    +    target_link_libraries(${executable} ${link_mode} LLVM)
       endif()

       explicit_llvm_config(${executable} ${link_components})
    @@ -99,7 +105,7 @@

       llvm_map_components_to_libnames(LIBRARIES ${link_components})
       get_target_property(t ${executable} TYPE)
    -  if(t STREQUAL "STATIC_LIBRARY")
    +  if(t STREQUAL "STATIC_LIBRARY" OR t STREQUAL "INTERFACE_LIBRARY")
         target_link_libraries(${executable} INTERFACE ${LIBRARIES})
       elseif(t STREQUAL "EXECUTABLE" OR t STREQUAL "SHARED_LIBRARY" OR t STREQUAL "MODULE_LIBRARY")
         target_link_libraries(${executable} PRIVATE ${LIBRARIES})