Constant extents need to be inferred pre storage flattening
Consider an allocation that has a dynamic extent, but needs to have a
constant extent because it's stored in MemoryType::Register (e.g. see
the test). We take an upper bound in these cases to get a constant
allocation size. This PR changes things to take that upper bound
*before* storage flattening instead of after. This way the individual
per-dimension extents are all constant, instead of just their product.
If you do it after storage flattening then you get dynamic strides
within a constant-sized allocation, which is silly and not compatible
with hoisting values into registers anyway (because access is at
non-constant coords).

Also fixed the assumption that MemoryType::Register on the GPU means
that there must be a GPULanes loop.