Clarify debug logging in copy_to_device()
We currently always call copy_to_device() on buffers that need to be on device (with the understanding that it's a no-op if no copy is needed); if the `debug` feature is on, a naive reading might make someone think that needless copy-to-device operations are actually happening. This adds a bit of logging (debug mode only) to make it clearer whether the copy to device actually happened, or if it was skipped because host was not dirty.