Add an HL_MEMORY_LIMIT environment var for the autoscheduler
This bounds the maximum amount of memory a generated pipeline may
allocate. If you set it too low the autoscheduler's search can run out
of legal states. Increasing HL_BEAM_SIZE may help it get out of the
hole, or there may simply not be any schedule that hits that memory
limit that doesn't also do a ridiculous amount of recompute.

Environment variables aren't great, but we don't have another means to
pass config to the autoscheduler right now.